Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,7 +4,7 @@ int main ()
     float a,b;
     scanf("%.2f",&a);
     scanf("%.2f",&b);
-    int product = a*b ;
-    printf("Product: %d", product);
+    float product = a*b ;
+    printf("Product: %.2f", product);
 
 }
